+++ 5black/5week/sprav_q/routing_settings.cpp
@@ -3,6 +3,6 @@
 using namespace std;
 
 RoutingSettings::RoutingSettings(const Json::Dict& dict) {
-  bus_wait_time = dict.at("bus_wait_time").AsInt();
-  bus_velocity = static_cast<double>(dict.at("bus_velocity").AsInt()) * 1000 / 60;
+  bus_wait_time = dict.at("bus_wait_time").AsDouble();
+  bus_velocity = static_cast<double>(dict.at("bus_velocity").AsDouble()) * 1000 / 60;
 }
